chore: move layouts into appropriate dirs
