+++ apps/trading/components/layouts-inner/index.ts
@@ -1,0 +1,2 @@
+export { LayoutWithSky } from './layout-with-sky';
+export { LayoutWithGradient } from './layout-with-gradient';
