+++ apps/trading/components/layouts/index.ts
@@ -1,1 +1,2 @@
-export * from './layout-with-sidebar';
+export { LayoutWithSidebar } from './layout-with-sidebar';
+export { LayoutCentered } from './layout-centered';
