+++ apps/trading/pages/client-router.tsx
@@ -2,8 +2,8 @@ import type { RouteObject } from 'react-router-dom';
 import { Navigate, useRoutes } from 'react-router-dom';
 import { lazy, Suspense } from 'react';
 import { Loader, Splash } from '@vegaprotocol/ui-toolkit';
-import { LayoutWithSidebar } from '../components/layouts';
-import { LayoutCentered } from '../components/layouts/layout-centered';
+import { LayoutWithSidebar, LayoutCentered } from '../components/layouts';
+import { LayoutWithSky } from '../components/layouts-inner';
 import { Home } from '../client-pages/home';
 import { Liquidity } from '../client-pages/liquidity';
 import { MarketsPage } from '../client-pages/markets';
@@ -15,7 +15,6 @@ import { Transfer } from '../client-pages/transfer';
 import { Fees } from '../client-pages/fees';
 import { Rewards } from '../client-pages/rewards';
 import { Routes as AppRoutes } from '../lib/links';
-import { LayoutWithSky } from '../client-pages/referrals/layout';
 import { Referrals } from '../client-pages/referrals/referrals';
 import { ReferralStatistics } from '../client-pages/referrals/referral-statistics';
 import { ApplyCodeFormContainer } from '../client-pages/referrals/apply-code-form';
